+++ packages/apiops-design-system/src/testing/index.tsx
@@ -1,0 +1,1 @@
+export { CanvasSystemFixture } from "../react/index.js";